requirements.txt
beautifulsoup4==4.12.3
discord.py==2.4.0
langchain==0.3.7
langchain_community==0.3.7
langchain_core==0.3.18
langchain_groq==0.2.1
langchain_huggingface==0.1.2
Requests==2.32.3
huggingface-hub==0.25.2
